refactor: move test utility functions to __utils for better organization
