+++ packages/adapters/test/__utils.ts
@@ -4,7 +4,12 @@ import type { GenericParseOptions } from "@mojis/parsers";
 import type { type } from "arktype";
 import type { FallbackFn, PredicateFn } from "../src/builders/adapter-builder/types";
 import type { AnyVersionHandler, VersionHandler } from "../src/builders/version-builder/types";
-import type { AdapterContext, AdapterHandlerType, BuiltinParser, PossibleUrls } from "../src/global-types";
+import type {
+  AdapterContext,
+  AdapterHandlerType,
+  BuiltinParser,
+  PossibleUrls,
+} from "../src/global-types";
 import { createCache } from "@mojis/internal-utils";
 import { vi } from "vitest";
 
@@ -18,55 +23,6 @@ type HANDLER_MAP = {
     fallback?: ORIGINAL_HANDLERS[K]["fallback"];
   };
 };
-
-export type CreateAdapterVersionHandler<TConfig extends {
-  output: any;
-  params?: Record<string, any>;
-}> = VersionHandler<{
-  globalContext: AdapterContext;
-  fetchOptions: RequestInit;
-  cacheOptions: CacheOptions;
-  parser: BuiltinParser;
-  parserOptions: GenericParseOptions;
-  parserOutput: unknown;
-  outputSchema: type.Any;
-  urls: PossibleUrls;
-  transform: unknown;
-  aggregate: unknown;
-  output: TConfig["output"];
-}>;
-
-export interface TestBuiltAdapterHandlerParams {
-  adapterType: string;
-  handlers: [PredicateFn, AnyVersionHandler][];
-  outputSchema?: type.Any;
-  fallback?: FallbackFn<any>;
-}
-
-export interface TestAdapterHandler<TParams extends TestBuiltAdapterHandlerParams> {
-  adapterType: TParams["adapterType"];
-  handlers: TParams["handlers"];
-  outputSchema?: TParams["outputSchema"];
-  fallback?: FallbackFn<
-    TParams["outputSchema"] extends type.Any
-      ? TParams["outputSchema"]["infer"]
-      : any
-  >;
-}
-
-export type CreateAnyAdapterHandler<
-  TType extends string,
-  TConfig extends {
-    handlers: CreateAdapterVersionHandler<any>[];
-    outputSchema?: type.Any;
-    fallback?: any;
-  },
-> = TestAdapterHandler<{
-  adapterType: TType;
-  handlers: Array<[PredicateFn, TConfig["handlers"][number]]>;
-  outputSchema?: TConfig["outputSchema"];
-  fallback?: FallbackFn<TConfig["handlers"][number]["output"]>;
-}>;
 
 export function createMockHandlers(): HANDLER_MAP {
   return {
@@ -179,3 +135,52 @@ export function cleanupAdapterTest(): void {
   vi.resetModules();
   vi.clearAllMocks();
 }
+
+export type CreateAdapterVersionHandler<TConfig extends {
+  output: any;
+  params?: Record<string, any>;
+}> = VersionHandler<{
+  globalContext: AdapterContext;
+  fetchOptions: RequestInit;
+  cacheOptions: CacheOptions;
+  parser: BuiltinParser;
+  parserOptions: GenericParseOptions;
+  parserOutput: unknown;
+  outputSchema: type.Any;
+  urls: PossibleUrls;
+  transform: unknown;
+  aggregate: unknown;
+  output: TConfig["output"];
+}>;
+
+export interface TestBuiltAdapterHandlerParams {
+  adapterType: string;
+  handlers: [PredicateFn, AnyVersionHandler][];
+  outputSchema?: type.Any;
+  fallback?: FallbackFn<any>;
+}
+
+export interface TestAdapterHandler<TParams extends TestBuiltAdapterHandlerParams> {
+  adapterType: TParams["adapterType"];
+  handlers: TParams["handlers"];
+  outputSchema?: TParams["outputSchema"];
+  fallback?: FallbackFn<
+    TParams["outputSchema"] extends type.Any
+      ? TParams["outputSchema"]["infer"]
+      : any
+  >;
+}
+
+export type CreateAnyAdapterHandler<
+  TType extends string,
+  TConfig extends {
+    handlers: CreateAdapterVersionHandler<any>[];
+    outputSchema?: type.Any;
+    fallback?: any;
+  },
+> = TestAdapterHandler<{
+  adapterType: TType;
+  handlers: Array<[PredicateFn, TConfig["handlers"][number]]>;
+  outputSchema?: TConfig["outputSchema"];
+  fallback?: FallbackFn<TConfig["handlers"][number]["output"]>;
+}>;
