+++ packages/adapters/test/global-types.test-d.ts
