+++ packages/adapters/test/handlers/adapters/sequences.test.ts
@@ -2,7 +2,7 @@ import { HttpResponse, mockFetch } from "#msw-utils";
 import { sequences } from "@mojis/loomicode";
 import { afterEach, describe, expect, it } from "vitest";
 import { handler } from "../../../src/handlers/adapter/sequences";
-import { cleanupAdapterTest, setupAdapterTest } from "../../test-utils";
+import { cleanupAdapterTest, setupAdapterTest } from "../../__utils";
 
 describe("sequences adapter handler", () => {
   const mockContext = {
