+++ packages/adapters/test/handlers/adapters/unicode-names.test.ts
@@ -2,7 +2,7 @@ import { HttpResponse, mockFetch } from "#msw-utils";
 import { createCache } from "@mojis/internal-utils";
 import { afterEach, describe, expect, it } from "vitest";
 import { handler } from "../../../src/handlers/adapter/unicode-names";
-import { cleanupAdapterTest, setupAdapterTest } from "../../test-utils";
+import { cleanupAdapterTest, setupAdapterTest } from "../../__utils";
 
 describe("unicode-names adapter handler", () => {
   const mockContext = {
